Move data from v1/full to v1 non-full api for purchase track in SDK (#8345)
