+++ packages/discovery-provider/src/api/v1/models/access_gate.py
@@ -3,8 +3,24 @@ from flask_restx import fields
 from .common import ns
 from .extensions.models import OneOfModel, WildcardModel
 
-tip_gate = ns.model("tip_gate", {"tip_user_id": fields.Integer(required=True)})
-follow_gate = ns.model("follow_gate", {"follow_user_id": fields.Integer(required=True)})
+tip_gate = ns.model(
+    "tip_gate",
+    {
+        "tip_user_id": fields.Integer(
+            required=True,
+            description="Must tip the given user ID to unlock",
+        )
+    },
+)
+follow_gate = ns.model(
+    "follow_gate",
+    {
+        "follow_user_id": fields.Integer(
+            required=True,
+            description="Must follow the given user ID to unlock",
+        )
+    },
+)
 nft_collection = ns.model(
     "nft_collection",
     {
@@ -16,7 +32,14 @@ nft_collection = ns.model(
     },
 )
 nft_gate = ns.model(
-    "nft_gate", {"nft_collection": fields.Nested(nft_collection, required=True)}
+    "nft_gate",
+    {
+        "nft_collection": fields.Nested(
+            nft_collection,
+            required=True,
+            description="Must hold an NFT of the given collection to unlock",
+        )
+    },
 )
 
 wild_card_split = WildcardModel(
@@ -32,7 +55,14 @@ usdc_gate = ns.model(
     },
 )
 purchase_gate = ns.model(
-    "purchase_gate", {"usdc_purchase": fields.Nested(usdc_gate, required=True)}
+    "purchase_gate",
+    {
+        "usdc_purchase": fields.Nested(
+            usdc_gate,
+            required=True,
+            description="Must pay the total price and split to the given addresses to unlock",
+        )
+    },
 )
 
 access_gate = ns.add_model(
